requirements.txt
torch~=2.3.1+cu121
easyocr~=1.7.1
numpy~=1.26.3
matplotlib~=3.9.0
pillow~=10.2.0
torchvision~=0.18.1+cu121
tqdm~=4.66.4